[+] hw05: implement king's moves

+++ 05-bitboard/src/bitboard/bishop.cpp
@@ -5,7 +5,7 @@ namespace bitboard {
 
 namespace bishop {
 
-uint64_t get_bitmask(uint64_t pos) { return 1; }
+uint64_t get_bitmask(int pos) { return 1; }
 
 } // namespace bishop
